requirements.txt
langchain-openai
langchain
pandas
python-dotenv
tabulate